\name{cDock-package}
\alias{cDock-package}
\alias{cDock}
\docType{package}
\title{
\packageTitle{cDock}
}
\description{
\packageDescription{cDock}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cDock}
\packageIndices{cDock}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cDock}

Maintainer: \packageMaintainer{cDock}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
